Update to use hubot-slack 4.10.0
Thanks to the new @slack/client npm, slackhq/hubot-slack#360, and
slackhq/hubot-slack#363, the `SlackBot` will now emit a
`ReactionMessage` object for `reaction_added` and `reaction_removed`
events.

A future commit will repackage the logic so that it is no longer
middleware, but called as part of a `robot.listen` or `robot.react`
callback.