SlackClient, DataStore: Make async
The SlackClient interface was originally synchronous because the
underlying Slack RTM client and in-memory data store provided by Hubot
and node-slack-sdk presented a synchronous interface.

We're updating the code to an asynchronous interface so we can
substitute an out-of-process data store. This allows the use of the same
core logic between the Hubot plugin and the upcoming Slack app.